.abort XML parsing when invalid characters were found in the input stream  (acid3-70) .add helper methods to KEncodingDetector for reporting such errors .add unit test asserting some KEncodingDetector properties
kindly reviewed by Maksim Orlovich.

CCBUG: 156947

svn path=/trunk/KDE/kdelibs/; revision=978411

+++ khtml/xml/xml_tokenizer.h
@@ -93,8 +93,8 @@ public:
 
     bool fatalError( const QXmlParseException& exception );
 
-    unsigned long errorLine;
-    unsigned long errorCol;
+    int errorLine;
+    int errorCol;
 
 private:
     void pushNode( DOM::NodeImpl *node );
